Creating event processor class to check for events and send them to the server.
